# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin-Panel-Api",
        "description": "Laravel Api",
        "contact": {
            "name": "Hamed",
            "email": "h.akbarzadh@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/admin/product-category/show/{productCategory}": {
            "get": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Get Product Category Detail",
                "description": "Get Product Category Detail",
                "operationId": "Categoryshow",
                "parameters": [
                    {
                        "name": "productCategory",
                        "in": "path",
                        "description": "Id of Product Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product-category/store": {
            "post": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Store Product Category in DB",
                "description": "Store Product Category in DB",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "show_in_menu",
                                    "status",
                                    "tags",
                                    "parent_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Category name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Category description"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Category image (OPT)"
                                    },
                                    "show_in_menu": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "0 Or 1"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "0 Or 1"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "Parent ID Category"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product-category/update/{productCategory}": {
            "put": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Update Prorduct Category in DB",
                "description": "Update Prorduct Category in DB",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "productCategory",
                        "in": "path",
                        "description": "Id of Article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "show_in_menu",
                                    "status",
                                    "tags",
                                    "parent_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Category name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Category description"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Category image (OPT)"
                                    },
                                    "show_in_menu": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "0 Or 1"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "0 Or 1"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "Parent ID Category"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product-category/destroy/{productCategory}": {
            "delete": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Delete Product Category",
                "description": "Delete Product Category",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "productCategory",
                        "in": "path",
                        "description": "Id of Product Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product-category/parent/{productCategory}": {
            "get": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Get Product Category Parent Detail",
                "description": "Get Product Category Parent Detail",
                "operationId": "parent",
                "parameters": [
                    {
                        "name": "productCategory",
                        "in": "path",
                        "description": "Id of Product Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product-category/children/{productCategory}": {
            "get": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Get Product Category Children Detail",
                "description": "Get Product Category Children Detail",
                "operationId": "children",
                "parameters": [
                    {
                        "name": "productCategory",
                        "in": "path",
                        "description": "Id of Product Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product-category/get-category/{number}": {
            "get": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Get Any Product Categories Detail",
                "description": "Get Any Product Categories Detail",
                "operationId": "getCategory",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Id of Product Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product-category/status/{productCategory}": {
            "get": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Change Status",
                "description": "Change Status",
                "operationId": "status",
                "parameters": [
                    {
                        "name": "productCategory",
                        "in": "path",
                        "description": "Id of Product Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}